requirements.txt
aiofiles==24.1.0
aiohttp==3.11.12
beautifulsoup4==4.13.3
elasticsearch==8.17.1
mysql_connector_repackaged==0.3.1
numpy==2.2.2
pandas==2.2.3
pytest==8.1.1
redis==5.2.1
Requests==2.32.3
selenium==4.28.1
setuptools==75.1.0
Sphinx==8.1.3
webdriver_manager==4.0.2
lxml==4.6.3